Ignore Nothing/Null when looking for dependencies
When looking for a value of a certain type in the eclosing and the
parent class, values of type Nothing or Null should be ignored.

For example, ScalaTest 2.0 defines the nullary methods cancel and fail
of type Nothing, which prevent using a mixed in MacWire module in a
test because they match any type.