reorganize tests to make sure they do not depend on more than needed #49
Test cases that were depending on `util` have been moved to the `utilTest`
project (which obviously depend on `util`). The `tests` project does not
depend anymore on util, proving that some of macwire features do not require
this extra dependency.

Another change is that `util` is now a `provided` dependency of `macros`.