make forward reference in block ineligible
Do not allow the following:
```
def build: B = {
  val b = wire[B]
  val forwardA = A() // forward reference is ineligible
}
```
Note that even if that would work with lazy vals, this isn't allowed:
```
def build: B = {
  lazy val b = wire[B]
  lazy val forwardA = A()
}
```
As explained in the todo test-case, wiring forward references in a block
is probably an anti-pattern: wiring in a block is useful when you need to
construct a complex object and return it. You seldom need multiple wire
calls with forward references.

A warning is emitted when both a value is found and a forward reference as
this is likely to be a mistake, or at least a source of confusion.